refactor(wallet-app): improve hooks import
